merge : back into develop

See merge request s06-bigdata-rec-sub2/S06P22E202!336

Former-commit-id: 7d820af7dd68978a2115dbd36d1ceb8227d5c81e

+++ backend/backend-java/src/main/java/com/ssafy/db/repository/LikeGoodsRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface LikeGoodsRepository  extends JpaRepository<LikeGoods, Long> {
     @Query(value = "select GOODS_ID from LikeGoods where USER_ID = :USERID",nativeQuery = true)
-    List<Integer> findGoodsIDByUserId(int USERID);
+    List<Long> findGoodsIDByUserId(int USERID);
 
     @Query(value =  "SELECT EXISTS(SELECT * FROM LikeGoods WHERE GOODS_ID = :goodsId AND USER_ID = :userId)", nativeQuery = true)
     int existsByGoodsIDAndUserID(long goodsId,long userId);
